working on climate/landscape

+++ dirt/gridworld2d/geology.py
@@ -6,7 +6,7 @@ import jax.random as jrng
 import chex
 from dirt.gridworld2d.perlin_noise import perlin_noise
 
-from dirt.defaults import DEFAULT_FLOAT_DTYPE
+from dirt.constants import DEFAULT_FLOAT_DTYPE
 
 '''
 The total world size is set to be (x,y,h)
@@ -39,8 +39,8 @@ def fractal_noise(
     # build the coords
     x_scale = grid_unit_scale * world_size[0]
     y_scale = grid_unit_scale * world_size[1]
-    x = jnp.linspace(0, x_scale, world_size[0])
-    y = jnp.linspace(0, y_scale, world_size[1])
+    x = jnp.linspace(0, x_scale, world_size[0], dtype=dtype)
+    y = jnp.linspace(0, y_scale, world_size[1], dtype=dtype)
     xx, yy = jnp.meshgrid(x, y, indexing='ij')
     coords = jnp.stack([xx.ravel(), yy.ravel()], axis=-1)
     
@@ -49,8 +49,8 @@ def fractal_noise(
     keys = jrng.split(key, max_octaves)[:octaves]
 
     # Frequency and amplitude for each octave
-    frequencies = lacunarity ** jnp.arange(octaves)
-    amplitudes = persistence ** jnp.arange(octaves)
+    frequencies = (lacunarity ** jnp.arange(octaves)).astype(dtype)
+    amplitudes = (persistence ** jnp.arange(octaves)).astype(dtype)
 
     # Vectorize across octaves
     def octave_noise(key, frequency, amplitude):
@@ -59,5 +59,5 @@ def fractal_noise(
     
     octave_noises = jax.vmap(octave_noise, 
                 in_axes=(0, 0, 0))(keys, frequencies, amplitudes)
-
+    
     return jnp.sum(octave_noises, axis=0).reshape(world_size) * height_scale
